Modulariza lógica en múltiples módulos
