+++ js/elements.js
@@ -1,0 +1,130 @@
+export const getDOM = () => {
+  const allViews = document.querySelectorAll('.view');
+  const navButtons = document.getElementById('nav-buttons');
+  const el = {
+    loadingView: document.getElementById('loading-view'),
+    loginView: document.getElementById('login-view'),
+    mainContent: document.getElementById('main-content'),
+    mainMenu: document.getElementById('main-menu-view'),
+    setup: document.getElementById('setup-view'),
+    item: document.getElementById('item-view'),
+    summary: document.getElementById('summary-view'),
+    history: document.getElementById('history-view'),
+    historyDetail: document.getElementById('history-detail-view'),
+    manageItems: document.getElementById('manage-items-view'),
+    finished: document.getElementById('finished-view'),
+    orderFinished: document.getElementById('order-finished-view'),
+    selectInventory: document.getElementById('select-inventory-view'),
+    setupOrder: document.getElementById('setup-order-view'),
+    orderItem: document.getElementById('order-item-view'),
+    orderSummary: document.getElementById('order-summary-view'),
+    editOrder: document.getElementById('edit-order-view'),
+    editInventory: document.getElementById('edit-inventory-view'),
+    consumptionSetup: document.getElementById('consumption-setup-view'),
+    consumptionResult: document.getElementById('consumption-result-view'),
+    continueInventoryBtn: document.getElementById('continue-inventory-btn'),
+    continueOrderBtn: document.getElementById('continue-order-btn'),
+    startNewInventoryBtn: document.getElementById('start-new-inventory-btn'),
+    makeOrderBtn: document.getElementById('make-order-btn'),
+    consumptionReportBtn: document.getElementById('consumption-report-btn'),
+    historyBtn: document.getElementById('history-btn'),
+    manageItemsBtn: document.getElementById('manage-items-btn'),
+    inventoryDate: document.getElementById('inventory-date'),
+    inventoryTime: document.getElementById('inventory-time'),
+    startCountingBtn: document.getElementById('start-counting-btn'),
+    itemCounter: document.getElementById('item-counter'),
+    itemName: document.getElementById('item-name'),
+    itemQuantity: document.getElementById('item-quantity'),
+    naBtn: document.getElementById('na-btn'),
+    skipBtn: document.getElementById('skip-btn'),
+    nextBtn: document.getElementById('next-btn'),
+    summaryList: document.getElementById('summary-list'),
+    inventorySummarySearch: document.getElementById('inventory-summary-search'),
+    backToCountBtn: document.getElementById('back-to-count-btn'),
+    historyList: document.getElementById('history-list'),
+    backToMenuFromHistoryBtn: document.getElementById('back-to-menu-from-history-btn'),
+    historyDetailTitle: document.getElementById('history-detail-title'),
+    historyDetailList: document.getElementById('history-detail-list'),
+    historyDetailPdfBtn: document.getElementById('history-detail-pdf-btn'),
+    backToHistoryBtn: document.getElementById('back-to-history-btn'),
+    manageItemsList: document.getElementById('manage-items-list'),
+    manageItemsSearch: document.getElementById('manage-items-search'),
+    newItemName: document.getElementById('new-item-name'),
+    addItemBtn: document.getElementById('add-item-btn'),
+    backToMenuFromManageBtn: document.getElementById('back-to-menu-from-manage-btn'),
+    saveAndFinishBtn: document.getElementById('save-and-finish-btn'),
+    finalDownloadBtn: document.getElementById('final-download-btn'),
+    saveOrderBtn: document.getElementById('save-order-btn'),
+    orderFinishedDownloadPdfBtn: document.getElementById('order-finished-download-pdf-btn'),
+    globalSummaryBtn: document.getElementById('summary-btn'),
+    globalPdfBtn: document.getElementById('download-pdf-btn'),
+    saveAndExitBtn: document.getElementById('save-and-exit-btn'),
+    discardProgressBtn: document.getElementById('discard-progress-btn'),
+    saveItemOrderBtn: document.getElementById('save-item-order-btn'),
+    toast: document.getElementById('toast-notification'),
+    confirmDeleteModal: document.getElementById('confirm-delete-modal'),
+    confirmDeleteTitle: document.getElementById('confirm-delete-title'),
+    confirmDeleteMessage: document.getElementById('confirm-delete-message'),
+    cancelDeleteBtn: document.getElementById('cancel-delete-btn'),
+    confirmDeleteBtn: document.getElementById('confirm-delete-btn'),
+    selectInventoryList: document.getElementById('select-inventory-list'),
+    backToMenuFromSelectInvBtn: document.getElementById('back-to-menu-from-select-inv-btn'),
+    orderForDate: document.getElementById('order-for-date'),
+    backToSelectInvBtn: document.getElementById('back-to-select-inv-btn'),
+    startOrderingBtn: document.getElementById('start-ordering-btn'),
+    orderItemCounter: document.getElementById('order-item-counter'),
+    orderItemName: document.getElementById('order-item-name'),
+    orderItemStock: document.getElementById('order-item-stock'),
+    orderItemQuantity: document.getElementById('order-item-quantity'),
+    orderNoPedirBtn: document.getElementById('order-no-pedir-btn'),
+    orderSkipBtn: document.getElementById('order-skip-btn'),
+    orderNextBtn: document.getElementById('order-next-btn'),
+    orderSummaryTitle: document.getElementById('order-summary-title'),
+    orderSummaryList: document.getElementById('order-summary-list'),
+    orderSummarySearch: document.getElementById('order-summary-search'),
+    backToOrderBtn: document.getElementById('back-to-order-btn'),
+    editOrderTitle: document.getElementById('edit-order-title'),
+    editOrderList: document.getElementById('edit-order-list'),
+    backToHistoryFromEditBtn: document.getElementById('back-to-history-from-edit-btn'),
+    saveEditedOrderBtn: document.getElementById('save-edited-order-btn'),
+    editInventoryTitle: document.getElementById('edit-inventory-title'),
+    editInventoryList: document.getElementById('edit-inventory-list'),
+    backToHistoryFromEditInvBtn: document.getElementById('back-to-history-from-edit-inv-btn'),
+    saveEditedInventoryBtn: document.getElementById('save-edited-inventory-btn'),
+    backToMenuFromConsumptionBtn: document.getElementById('back-to-menu-from-consumption-btn'),
+    startInventorySelect: document.getElementById('start-inventory-select'),
+    endInventorySelect: document.getElementById('end-inventory-select'),
+    generateConsumptionReportBtn: document.getElementById('generate-consumption-report-btn'),
+    consumptionResultList: document.getElementById('consumption-result-list'),
+    backToConsumptionSetupBtn: document.getElementById('back-to-consumption-setup-btn'),
+    logoutBtn: document.getElementById('logout-btn'),
+    userEmail: document.getElementById('user-email'),
+    loginError: document.getElementById('login-error-message'),
+    authTitle: document.getElementById('auth-title'),
+    emailInput: document.getElementById('email-input'),
+    passwordInput: document.getElementById('password-input'),
+    authActionBtn: document.getElementById('auth-action-btn'),
+    authToggleText: document.getElementById('auth-toggle-text'),
+    authToggleBtn: document.getElementById('auth-toggle-btn'),
+    orderMethodView: document.getElementById('order-method-view'),
+    orderMethodManualBtn: document.getElementById('order-method-manual-btn'),
+    orderMethodRemitoBtn: document.getElementById('order-method-remito-btn'),
+    backToSetupOrderBtn: document.getElementById('back-to-setup-order-btn'),
+    remitoUploadView: document.getElementById('remito-upload-view'),
+    backToOrderMethodBtn: document.getElementById('back-to-order-method-btn'),
+    remitoUploadCreation: document.getElementById('remito-upload-creation'),
+    remitoPreviewContainer: document.getElementById('remito-preview-container'),
+    analyzeCreationRemitoBtn: document.getElementById('analyze-creation-remito-btn'),
+    processingCreationLoader: document.getElementById('processing-creation-loader'),
+    remitoConfirmView: document.getElementById('remito-confirm-view'),
+    backToRemitoUploadBtn: document.getElementById('back-to-remito-upload-btn'),
+    remitoConfirmList: document.getElementById('remito-confirm-list'),
+    confirmRemitoDataBtn: document.getElementById('confirm-remito-data-btn'),
+    addRemitoItemSelect: document.getElementById('add-remito-item-select'),
+    addRemitoItemQuantity: document.getElementById('add-remito-item-quantity'),
+    addRemitoItemBtn: document.getElementById('add-remito-item-btn'),
+    remitoItemCount: document.getElementById('remito-item-count'),
+    remitoBultoCount: document.getElementById('remito-bulto-count'),
+  };
+  return { allViews, navButtons, el };
+};
