+++ styles.css
@@ -1,0 +1,15 @@
+body { font-family: 'Inter', sans-serif; }
+::-webkit-scrollbar { width: 8px; }
+::-webkit-scrollbar-track { background: #1f2937; }
+::-webkit-scrollbar-thumb { background: #4b5563; border-radius: 10px; }
+input[type=number]::-webkit-inner-spin-button,
+input[type=number]::-webkit-outer-spin-button { -webkit-appearance: none; margin: 0; }
+input[type=number] { -moz-appearance: textfield; }
+.view { transition: opacity 0.3s ease-in-out; }
+details > summary { list-style: none; }
+details > summary::-webkit-details-marker { display: none; }
+.sticky-header { position: sticky; top: 0; background-color: #374151; }
+.sticky-footer { position: sticky; bottom: 0; background-color: #1f2937; }
+.drag-handle { cursor: grab; }
+.dragging { opacity: 0.5; background: #4f46e5; }
+.drag-over { border-top: 2px solid #a5b4fc; }
